Updates by module7-task2
