+++ src/components/offers-container/offers-container.connect.jsx
@@ -1,10 +1,11 @@
 import {connect} from 'react-redux';
+import NameSpace from '../../reducer/name-space';
 import OffersContainer from './offers-container';
 
 const mapStateToProps = (state) => ({
-  offers: state.offers,
-  city: state.city,
-  offersByCity: state.offersByCity
+  offers: state[NameSpace.DATA].offers,
+  city: state[NameSpace.DATA].city,
+  offersByCity: state[NameSpace.DATA].offersByCity
 });
 
 export default connect(mapStateToProps)(OffersContainer);
