+++ src/components/offers-container/offers-container.jsx
@@ -15,7 +15,7 @@ const OffersContainer = React.memo(function OffersContainer(props) {
     offersByCity,
   } = props;
 
-  const offers = offersByCity.length > 0 ? offersByCity : props.offers;
+   const offers = offersByCity.length > 0 ? offersByCity : props.offers;
 
   return (
     <div className="cities__places-container container">
