+++ src/components/offers-list/offers-list.connect.jsx
@@ -1,8 +1,10 @@
 import {connect} from 'react-redux';
+import NameSpace from '../../reducer/name-space';
 import OffersList from './offers-list';
 
 export const mapStateToProps = (state) => ({
-  offersByCity: state.offersByCity
+  offers: state[NameSpace.DATA].offers,
+  offersByCity: state[NameSpace.DATA].offersByCity
 });
 
 export default connect(mapStateToProps)(OffersList);
