+++ src/components/offers-list/offers-list.jsx
@@ -1,9 +1,13 @@
 import Offer from "../offer/offer.connect";
 import {OfferProperties} from "../../proptypes/properties";
 
-const OffersList = ({offersByCity}) => {
+const OffersList = (props) => {
 
-  return offersByCity.map((offer, index) =>
+  const {offersByCity} = props;
+
+  const offers = offersByCity.length > 0 ? offersByCity : props.offers;
+    
+  return offers.map((offer, index) =>
     <Offer
       key={`offer-${index}`}
       {...offer}
