+++ src/consts.js
@@ -4,3 +4,7 @@ export const FilterType = {
   HIGH_TO_LOW: `Price: high to low`,
   TOP_RATED: `Top rated first`,
 };
+
+export const ErrorType = {
+  UNAUTHORIZED: 401
+}
